Adding initial revision of sDashboard V2.0 (aka 'Salt&Pepper')
Features Added
++++++++++++++++++++++++++++

1) Replaced Flot Charts with Flotr2 charting library
2) Fixed the maximize issue, the widget now maximize to full screen
3) Added code to dispatch sorted widget definitions array when widget
are rearranged
4) Added example to add an exisiting dom element as a widget by passing
in the element using a jquery selector.
5) Added examples for bar,line,pie, and bubble charts. All charts
supported be flot charts are support by the dashboard framwork now
6) Fixed the issue which resizing of chart widgets, earlier we were
maintaining two version of chart, one in minimized state and one in
maximized state and we were toggling between it when maximized or
minimized. This release fix that issue and make its much simpler buy
just redrawing the chart again.

Low level code changes:
7) Refactored unsed css
8)Removed isADonut option on widget definition as its not longer
supported as we switched to flort2 chart
9) Add new options for widget definitions
10) Instead of have a chart type = bar, pie, line etc.. refactored the
code to be more generic, i.e. if we want to render a chart set the
widgetType to chart and based on the options passed, the chart will be
rendered accordingly.